Create Strings_Making_Anagrams.cpp
Given two strings, a and b, that may or may not be of the same length, determine the minimum number of character deletions required to make and anagrams. Any characters can be deleted from either of the strings. 

Input : str1 = "bcadeh" str2 = "hea"
Output: 3
Input : str1 = "cddgk" str2 = "gcd"
Output: 2