Create String_Alternating_Characters.cpp
You are given a string containing characters A and B only. Your task is to change it into a string such that there are no matching adjacent characters. To do this, you are allowed to delete zero or more characters in the string. Your task is to find the minimum number of required deletions.
Sample Input:
5
AAAA
BBBBB
ABABABAB
BABABA
AAABBB
Sample Output:
3
4
0
0
4